Adjust style on hover
When using this on a tablet, these buttons would keep focus and it would
override the "selected" style, making it so that they were white instead
of blue. This was a bit confusing, so I used a style that doesn't
require a change in background.

I also noticed a few places where a style that should have only applied
to #floating-controls sub-elements was actually applying to all buttons.